+++ src/main/kotlin/com/kotcrab/ghidra/rest/mapper/FunctionMapper.kt
@@ -19,6 +19,7 @@ class FunctionMapper {
           parameters = function.parameters.map {
             ApiFunction.Parameter(
               ordinal = it.ordinal,
+              name = it.name,
               dataTypePathName = it.dataType.pathName
             )
           },