+++ CHANGES.md
@@ -1,3 +1,8 @@
+#### Version 2.1
+
+- Added `function.parameter[].name` field
+- Added `returnTypePathName` and `parameters` fields to the function prototype types
+
 #### Version 2
 
 - Breaking change: `symbol.offset` renamed to `symbol.address`